fix: git-tb check command
fix 'git_toolbox_all' function to properly traverse the repositories
and refactor a litle the function. Bear in mind the script is
executed with the 'set -e' option, which impact onto how to retrieve
the exit-code and how to handle some statements which could fails
but it were expected. Add --continue and --verbose arguments:
  - when --continue is indicated, it does not stop when the
    subcommand fails and continue traversing.
  - when --verbose is indicated, always print out an informative
    message indicating the directory that is being traversed.

fix 'git_toolbox_check__helper' to consider the 'set -e' flag.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>