fix: error message in tmux
When we run tmux, we are getting the error
'which -bash' when we were expecting something
like 'which bash'; this is an argument passed
and probably related on how tmux invoke the
shell.

This silent the error message (bad action) but
we do not have other way on this case how to
operate.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>